Added Accordion, Alert, AppLaunched, Card, Checkbox
