+++ components/accordion/accordion.md
@@ -1,0 +1,15 @@
+# accordion
+
+```javascript
+<Accordion>
+	<Panel expanded summary={'Panel 1. Merge'}>
+		{'Merge is a revolution'}
+	</Panel>
+	<Panel summary={'Panel 2. Merge'}>
+		{'Merge is a revolution'}
+	</Panel>
+	<Panel summary={'Panel 3. Merge'}>
+		{'Merge is a revolution'}
+	</Panel>
+</Accordion>
+```
