+++ components/accordion/panel/index.jsx
@@ -1,0 +1,3 @@
+import Panel from "./panel.jsx";
+
+export { Panel as default }
