+++ components/accordion/panel/panel.md
@@ -1,0 +1,10 @@
+# Panel
+
+```javascript
+<Panel
+	expanded
+	summary={"Panel 1. Merge"}
+>
+	{"Merge is a revolution"}
+</Panel>
+```
