+++ components/accordion/panel/presets/0-default.json
@@ -1,0 +1,13 @@
+{
+	"rootId": "accordionPanel3703996384",
+	"elements": {
+	  "accordionPanel3703996384": {
+		"name": "Panel",
+		"props": {
+			"expanded": true,
+			"summary": "Panel 1. Merge",
+			"children": "Merge is a revolution"
+		}
+	  }
+	}
+  }
