+++ components/alert/alert.md
@@ -1,0 +1,11 @@
+# Alert
+
+```javascript
+<Alert
+dismissible
+labels={{
+heading: 'Logged in as John Smith (johnsmith@acme.com).',
+headingLink: 'Log out',
+}}
+/>
+```
