+++ components/app-launcher/section/section.jsx
@@ -19,14 +19,14 @@ import isFunction from 'lodash.isfunction';
 import classNames from 'classnames';
 
 // This component's `checkProps` which issues warnings to developers about properties when in development mode (similar to React's built in development tools)
-import checkProps from './check-props';
-import componentDoc from './docs.json';
+import checkProps from '../check-props';
+import componentDoc from '../docs.json';
 
 // ## Children
-import Button from '../button';
+import Button from '../../button';
 
 // ## Constants
-import { APP_LAUNCHER_SECTION } from '../../utilities/constants';
+import { APP_LAUNCHER_SECTION } from '../../../utilities/constants';
 
 const defaultProps = {
 	assistiveText: {
