+++ components/app-launcher/section/section.md
@@ -1,0 +1,17 @@
+# Section
+
+```javascript
+<Section title="Tile Section">
+	<Tile
+		title="Marketing Cloud"
+		iconText="MC"
+		description="Send emails, track emails, read emails! Emails!"
+	/>
+	<Tile
+		title="Call Center"
+		description="The key to call center and contact center is not to use too many words!"
+		descriptionHeading="Call Center"
+		iconText="CC"
+	/>
+</Section>
+```
