+++ components/app-launcher/tile/tile.md
@@ -1,0 +1,10 @@
+# Tile
+
+```javascript
+<Tile
+	title="Call Center"
+	description="The key to call center and contact center is not to use too many words!"
+	descriptionHeading="Call Center"
+	iconText="CC"
+/>
+```
