+++ components/button/presets/0-default.jsx
@@ -1,4 +1,5 @@
+/* eslint-disable filenames/match-regex */
 import * as React from 'react';
 import Button from '../button.jsx';
 
-export default <Button label="merge!" />;
+export default (<Button label="merge!" />);
