+++ components/card/presets/0-default.json
@@ -1,0 +1,10 @@
+{
+	"rootId": "card3703996384",
+	"elements": {
+	  "card3703996384": {
+		"name": "Card",
+		"props": {
+		}
+	  }
+	}
+  }
