+++ components/checkbox/presets/0-default.json
@@ -1,0 +1,10 @@
+{
+	"rootId": "checkbox3703996384",
+	"elements": {
+	  "checkbox3703996384": {
+		"name": "Checkbox",
+		"props": {
+		}
+	  }
+	}
+  }
