+++ package.json
@@ -132,7 +132,7 @@
 		"@storybook/addon-storyshots": "^3.4.10",
 		"@storybook/addons": "^3.4.10",
 		"@storybook/react": "^3.4.10",
-		"@uxpin/merge-cli": "0.0.20",
+		"@uxpin/merge-cli": "^1.0.1",
 		"async": "^2.0.0-rc.5",
 		"babel-core": "^7.0.0-bridge.0",
 		"babel-eslint": "^8.2.3",
