+++ uxpin.config.js
@@ -5,11 +5,20 @@ module.exports = {
         name: 'General',
         include: [
 					'components/button/button.jsx',
-					'components/checkbox/checkbox.jsx',
-					'components/card/card.jsx',
         ]
       },
     ]
   },
   name: 'Salesforce Lightning Design System'
 };
+
+/*
+					'components/checkbox/checkbox.jsx',
+					'components/card/card.jsx',
+					'components/accordion/accordion.jsx',
+					'components/accordion/panel/panel.jsx',
+					'components/alert/alert.jsx',
+					'components/app-launcher/app-launcher.jsx',
+					'components/app-launcher/section/section.jsx',
+					'components/app-launcher/tile/tile.jsx',
+*/
